wfsfintech/backend/requirements.txt
fastapi
uvicorn
numpy
pandas
scikit-learn
xgboost
hmmlearn
pydantic
python-jose[cryptography]
yfinance
py_vollib
